prices as int and some code cleanup

+++ tests/PlayerInterface.java
@@ -9,5 +9,5 @@ public interface PlayerInterface {
 	public void setPosition(PositionMock pos) throws InvalidPosition;
 	public PositionMock getPosition();
 	public String getPositionName();
-	public Double getPrice();
+	public Integer getPrice();
 }
